+++ src/gui/App/hooks/useCardHandler.ts
@@ -75,7 +75,7 @@ export default (props:Props) => {
 			const stackIndex = findStackIndex(draft, event.stackId);
 			draft.stacks[stackIndex].cards.push({
 				id: newCardId,
-				title: 'New card',
+				title: '',
 				body: '',
 				is_todo: 0,
 				todo_completed: 0,